*) Code from alex_speed - Changed : RealmCheck and DistanceCheck in - GroundAssist.cs and moved the RealmCheck in assist.cs *) Code from SmallHorse - Changed: The design of the property system to streamline it *) Code from Tiv - Bugfix: Fixed an issue where in very rare cases the client sends 0xA9 packet before sending S<=C 0xE8 player world initialize. *) Code from Malkavien - Bugfix: Magical items delves now show again correctly *) Code from Biceps - Bugfix: Nant buildfile works again *) Code from Overdriven  - Changed: Wardens are now granted 1.8 spec points per level.  - Changed: Wardens are now able to specialize in Shield.  - Changed: Valewalkers now have access to Protect 1 at level 15.  - Changed: Valewalkers now have access to Intercept at level 19.  - Changed: Valewalkers now have access to Protect 2 at level 23.  - Changed: Valewalkers now have access to Protect 3 at level 32.  - Changed: Valkyrie's can now use large shields  - Changed: Hunters/Rangers/Scouts have been given SafeFall 1 *) Code from Me - Bugfix: You will no longer see "cancel all conc effects" twice when a player leaves the game. - New: Merged my Area Capture system with the DOL Keep system, with all it's functionality. - Removed: The speed hack detection from the ping request on the advice of Tiv because this is triggered by lag also

+++ GameServerScripts/gmcommands/keep.cs
@@ -17,11 +17,13 @@
  *
  */
 using System;
-using DOL.GS.PacketHandler;
-using DOL.Database;
-using DOL.AI.Brain;
 using System.Collections;
 using System.Collections.Specialized;
+
+using DOL.AI.Brain;
+using DOL.Database;
+using DOL.GS.Keeps;
+using DOL.GS.PacketHandler;
 
 namespace DOL.GS.Scripts
 {
@@ -1915,10 +1917,10 @@ namespace DOL.GS.Scripts
 						client.Out.SendMessage("You must create a keep first",eChatType.CT_System,eChatLoc.CL_SystemWindow);
 						return 1;
 					}
-					int keepLevel = 0;
+					byte keepLevel = 0;
 					try
 					{
-						keepLevel = Convert.ToInt32(args[2]);
+						keepLevel = Convert.ToByte(args[2]);
 					}
 					catch
 					{
